[REM-103] refactor: update login_required import paths
- Updated import paths for login_required decorator from web.custom to web.decorators in multiple files.
- Added new web.types module for type definitions.
- Introduced login_wrap.py for login_required decorator implementation.
