+++ web/routes/bot/__init__.py
@@ -21,7 +21,7 @@ from quart import (
 )
 from quart import current_app as app
 
-from web.custom import login_required
+from web.decorators import login_required
 from web.models import BotsCrawJUD
 
 from ...forms import BotForm
