+++ web/routes/config/admin/__init__.py
@@ -9,7 +9,7 @@ from flask_sqlalchemy import SQLAlchemy
 from quart import Blueprint, Response, abort, flash, make_response, redirect, render_template, session, url_for
 from quart import current_app as app
 
-from web.custom import login_required
+from web.decorators import login_required
 from web.forms import UserForm, UserFormEdit
 from web.models import LicensesUsers, SuperUser, Users
