+++ web/routes/config/superSu/bots.py
@@ -3,7 +3,7 @@
 import quart_flask_patch  # noqa: F401
 from quart import Response, abort, make_response, render_template
 
-from web.custom import login_required
+from web.decorators import login_required
 
 from ....decorators import check_privilegies
 from . import supersu
