+++ web/routes/config/user/__init__.py
@@ -9,7 +9,7 @@ import pathlib
 import quart_flask_patch  # noqa: F401
 from quart import Blueprint, Response, make_response, render_template
 
-from web.custom import login_required
+from web.decorators import login_required
 
 path_template = os.path.join(pathlib.Path(__file__).parent.resolve(), "templates")
 usr = Blueprint("usr", __name__, template_folder=path_template)
