+++ web/routes/execution/schedules.py
@@ -6,7 +6,7 @@ from quart import Response, abort, make_response, render_template, session
 from quart import current_app as app
 from sqlalchemy.orm import aliased
 
-from web.custom import login_required
+from web.decorators import login_required
 from web.models import ScheduleModel, SuperUser, Users, admins
 
 from . import exe
